refactor android user attributes
