+++ packages/amplify_auth_cognito/android/src/main/kotlin/com/amazonaws/amplify/amplify_auth_cognito/utils/UserAttributeValidation.kt
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ * or in the "license" file accompanying this file. This file is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
+package com.amazonaws.amplify.amplify_auth_cognito.utils
+
+import com.amazonaws.amplify.amplify_core.exception.ExceptionMessages
+import com.amazonaws.amplify.amplify_core.exception.InvalidRequestException
+
+fun validateUserAttribute(attribute: HashMap<*, *>, validationErrorMessage: String) {
+    if (!attribute.containsKey("value")) {
+        throw InvalidRequestException(validationErrorMessage, ExceptionMessages.missingAttribute.format("value"))
+    } else if (!attribute.containsKey("userAttributeKey")) {
+        throw InvalidRequestException(validationErrorMessage, ExceptionMessages.missingAttribute.format("userAttributeKey"))
+    } else if (attribute["value"] !is String) {
+        // Android SDK expects a string for user attr values, regardless of the configuration in cognito
+        throw InvalidRequestException(validationErrorMessage, "Attribute value is not a String.")
+    }
+}
